Move commit report builders to their own package and delete truncate.
