+++ commit/plugin.go
@@ -18,6 +18,7 @@ import (
 	"github.com/smartcontractkit/chainlink-common/pkg/merklemulti"
 	"github.com/smartcontractkit/chainlink-common/pkg/services"
 
+	"github.com/smartcontractkit/chainlink-ccip/commit/builder"
 	"github.com/smartcontractkit/chainlink-ccip/commit/chainfee"
 	"github.com/smartcontractkit/chainlink-ccip/commit/committypes"
 	"github.com/smartcontractkit/chainlink-ccip/commit/merkleroot"
@@ -48,7 +49,7 @@ type Plugin struct {
 	ccipReader        readerpkg.CCIPReader
 	tokenPricesReader readerpkg.PriceReader
 	reportCodec       cciptypes.CommitPluginCodec
-	reportBuilder     reportBuilderFunc
+	reportBuilder     builder.ReportBuilderFunc
 	// Don't use this logger directly but rather through logutil\.WithContextValues where possible
 	lggr                logger.Logger
 	homeChain           reader.HomeChain
@@ -83,6 +84,7 @@ func NewPlugin(
 	reportingCfg ocr3types.ReportingPluginConfig,
 	reporter metrics.Reporter,
 	addressCodec cciptypes.AddressCodec,
+	reportBuilder builder.ReportBuilderFunc,
 ) *Plugin {
 	lggr.Infow("creating new plugin instance", "p2pID", oracleIDToP2pID[reportingCfg.OracleID])
 
@@ -162,21 +164,6 @@ func NewPlugin(
 		reportingCfg.F,
 		reporter,
 	)
-
-	var reportBuilder reportBuilderFunc
-
-	// These options were added to allow for more flexibility around report building. For example Solana
-	// only supports a single merkle root per report.
-	maxRoots := offchainCfg.MaxMerkleRootsPerReport
-	if !offchainCfg.RMNEnabled && maxRoots != 0 {
-		if offchainCfg.MultipleReportsEnabled {
-			reportBuilder = buildMultipleReports
-		} else {
-			reportBuilder = buildTruncatedReport
-		}
-	} else {
-		reportBuilder = buildStandardReport
-	}
 
 	return &Plugin{
 		donID:               donID,
